// packages/agora-edu-core/src/stores/ui/common/layout/type.ts
export type Bounds = {
  left: number;
  top: number;
  right: number;
  bottom: number;
};

export type Margin = {
  top: number;
};

/**
 * 尺寸信息
 */
export type Dimensions = { width: number; height: number };

/**
 * 位置坐标点
 */
export type Point = { x: number; y: number };

/**
 * 缩放范围
 */
export type ResizeBounds = {
  minWidth: number;
  minHeight: number;
  maxWidth: number;
  maxHeight: number;
};